# packages/api-contracts/src/generated/types/repoStructure.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeMapAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CodeNode } from "./codeNode";

export interface RepoStructure {
  repoName: string;
  rootNode: CodeNode;
  fileCount: number;
  languages: string[];
  entryPoints?: string[];
  totalSize?: number;
  sourceUrl?: string;
  branch?: string;
}