# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CodeMapAI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /repo/fetch:
    post:
      operationId: fetchRepo
      tags: [repo]
      summary: Fetch repository structure
      description: Fetches and parses the structure of a GitHub repository or processes uploaded ZIP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FetchRepoRequest"
      responses:
        "200":
          description: Repository structure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepoStructure"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /repo/upload:
    post:
      operationId: uploadZip
      tags: [repo]
      summary: Upload ZIP file
      description: Upload a ZIP file containing a codebase
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: Repository structure extracted from ZIP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepoStructure"
        "400":
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    FetchRepoRequest:
      type: object
      properties:
        url:
          type: string
          description: GitHub repository URL
        branch:
          type: string
          description: Branch to fetch (defaults to main/master)
      required:
        - url
    CodeNode:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [root, folder, file, function, class, export]
        path:
          type: string
        children:
          type: array
          items:
            $ref: "#/components/schemas/CodeNode"
        language:
          type: string
        size:
          type: integer
        complexity:
          type: string
          enum: [low, medium, high]
        isEntryPoint:
          type: boolean
        content:
          type: string
          description: Optional source content for small text files
      required:
        - id
        - name
        - type
        - path
    RepoStructure:
      type: object
      properties:
        repoName:
          type: string
        rootNode:
          $ref: "#/components/schemas/CodeNode"
        fileCount:
          type: integer
        languages:
          type: array
          items:
            type: string
        entryPoints:
          type: array
          items:
            type: string
        totalSize:
          type: integer
        sourceUrl:
          type: string
          format: uri
        branch:
          type: string
      required:
        - repoName
        - rootNode
        - fileCount
        - languages
...